pyproject.toml
[project]
name = "msc-thesis-db"
version = "0.1.0"
description = "Prototype for a database of MSc thesis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2>=2.9.10",
    "pydantic-config>=0.3.0",
    "pydantic-string-url>=1.0.2",
    "sqlmodel>=0.0.24",
]

[tool.ruff]
line-length = 120

lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
lint.ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.lint.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = ["app"]

[tool.ruff.lint.flake8-bugbear]
# Allow default arguments like, e.g., `data: List[str] = fastapi.Query(None)`.
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]
"**/{tests,docs}/*" = ["E402"]

[tool.pyright]
include = ["app/", "scripts/"]
venv = ".venv"
venvPath = "."

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupysql>=0.11.1",
    "polars>=1.27.1",
]
